Addition of the fckEditor as an integrated option for the advanced_editor Initial setup of updates which includes bulk of required new files and changes.
Version 2.0 Final Candidate of fckEditor
Includes an integrated version of the mcpuk image browser which has been modified to configure the image browser settings based on the GL config.
The mcpuk browser is an additional component but is a more advanced plugin

Image library base directory path defined in the main GL config.php but will be by default /images/library.