Fixed issue of 2 many file locations added by template class to check for template files
For issue #855.

CTL_setTemplateRoot before was set back when child themes support was added (not sure which Geeklog version) as a hook to run as a template preprocessor and add any theme default directories for child themes. This was fine for Geeklog Core but could create issues for plugins as the function could not tell the difference and would add in theme and theme_default root dir locations for plugins (among other custom folders).

For backwards compatibility the hook and CTL_setTemplateRoot has been left in place (but modified to perform better) since some plugins still set the template class directly or may use COM_newTemplate without CTL_core_templatePath or CTL_plugin_templatePath.

**CTL_setTemplateRoot has now been depreciated and will disappear as of Geeklog 3.0.0. This means plugins CANNOT set the template class directly like below. They will be required to use COM_newTemplate and either CTL_plugin_templatePath or CTL_core_templatePath.**

Now for COM_newTemplate, if root is passed as a single directory (not an array) it is assumed that Geeklog Core is setting the template or an old style plugin and will leave the hook in place.

If root is an array it is assumed that either CTL_core_templatePath or CTL_plugin_templatePath is being used and the hook will be removed. For more info see template class set_root function, CTL_setTemplateRoot, CTL_core_templatePath, CTL_plugin_templatePath.

Also fixed a few spots where COM_newTemplate was not being used in Geeklog Core

The template class was being set directly like:

`$T = new Template($_DLM_CONF['path_layout']);`