Removed Setting of HomePage Link for Facebook OAuth
The user link permission gives a link to the person's Timeline. The problem is the link will only resolve if the person clicking the link is logged into Facebook and is a friend of the person whose profile is being viewed. So therefore it is no good for our user profile homepage link.

For issue #881