Added  PLG_getCSSClasses so css classes do not need to be hardcoded in code and can be theme independent
For Feature #903

A new theme/plugin function called PLG_getCSSClasses has been added. It works with core and plugins and allows us to remove hardcoding of css classes in code. You pass in an id and type. The type is a plugin or core type (like article) and basically specifies the location to look for the css classes. If it is a core type it will check the current theme functions.php for the proper function. If it is a plugin type it will check the plugin theme templates for funcitons.php and the proper function

Id is the name of the location like 'core-list-new' (which is the What's New Block List).