Update devel-db-update.php
This update includes the session table changes and user table changes for the new session class. Unfortantly these changes have to be applied manually since this script requires the user to be logged in to use it. and they obviously cannot be logged in until the table changes have happened