Fix for Usernames
For #931. For Usernames across local and remote services are required to be unique with no trailing spaces and case and accent insensitive. This should not demped on the database collation.

Also fix dupplicate email field showing up for denim three and modern curve theme. Also fixed maxlength on a few of the fields in the user editor.