Fixed Comment Report Abuse
To fix #1034 and in part for #1023

PLG_commentEnabled now returns the comment code to theplugin item. This is based not only the code of the actual plugin item but the access the user has to the item. This means things can happen for comments that are not only enabled but closed as well (like Report Abuse)