Fixed searching for Articles and Article Comments
Fix for #1044. Now results are returned just for articles and article comments are included when comments are just selected. Issue was the type was mislabled so search was looking for 'stories' instead of 'article'.

Also moved custom article search code from search class to lib-article. Articles is now treated like a plugin for searches.