Returned LANG variables from Template Class now exactly the same as found in language files
Initially htmlspecialchars was used on the return language variable to convert HTML into entities. Not sure why this is done as some language variables have HTML tags in them for formatting reasons.  When these language variables are used in code htmlspecialchars is not used.

If language variables need to display actual HTML tags then they should already use HTML entities and not need to be converted.